+++ src/test/java/de/is24/deadcode4j/analyzer/hibernateannotations/EntityWithGeneratedValue.java
@@ -1,0 +1,9 @@
+package de.is24.deadcode4j.analyzer.hibernateannotations;
+
+import javax.persistence.GeneratedValue;
+
+@SuppressWarnings("UnusedDeclaration")
+public class EntityWithGeneratedValue {
+    @GeneratedValue(generator = "generatorOne")
+    private String id;
+}